+++ src/cpx_io/cpx_system/cpx_ap/cpx_ap.py
@@ -118,7 +118,7 @@ class CpxAp(CpxBase):
             # if correct apdd exists in folder, use it!
             if apdd_name in apdds:
                 with open(
-                    self._apdd_path + "/" + apdd_name, "r", encoding="ascii"
+                    self._apdd_path + "/" + apdd_name, "r", encoding="utf-8"
                 ) as f:
                     module_apdd = json.load(f)
                 Logging.logger.debug(
@@ -435,25 +435,25 @@ class CpxAp(CpxBase):
         :rtype: dict"""
         reg = self.read_reg_data(self.diagnosis_register, length=2)
         diagnosis_keys = [
-            "device available",
-            "current",
-            "voltage",
-            "temperature",
+            "Device available",
+            "Current",
+            "Voltage",
+            "Temperature",
             "reserved",
-            "motion",
-            "configuration/parameters",
-            "monitoring",
-            "communication",
+            "Movement",
+            "Configuration / Parameters",
+            "Monitoring",
+            "Communication",
             "Safety",
-            "internal hardware",
-            "software",
-            "maintenance",
-            "miscellaneous",
+            "Internal Hardware",
+            "Software",
+            "Maintenance",
+            "Misc",
             "reserved(14)",
             "reserved(15)",
-            "external device",
-            "safety",
-            "encoder",
+            "External Device",
+            "Security",
+            "Encoder",
         ]
         # the rest of the bits are "reserved" and therefore trunctuated
         diagnoisis_dict = {
@@ -655,7 +655,7 @@ class CpxAp(CpxBase):
                 + fw_version.replace(".", "-")
                 + ".json"
             )
-            with open(output_file_path, "w", encoding="ascii") as f:
+            with open(output_file_path, "w", encoding="utf-8") as f:
                 f.write(json.dumps(json_data, indent=4))
             Logging.logger.debug(f"JSON data has been written to: {output_file_path}")
             return json_data