+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Checker script to read in APDDs and check if they are compatible with the library
 - Added type hint for ap modules property to enable autocompletion.
 - Added `int_to_boollist()` to utils.boollist
 
 ### Changed
+- CPX-AP now interprets apdds with utf-8 instead of ascii encoding
 - `CpxAp.read_global_diagnosis_state()` now returns dict according to "module diagnostics state" (see datasheet CPX-AP-\*-EP-\*) 
 
 ### Fixed